Refactor parse testing out
